Break apart request types to allow separate jwt reqs
