--- client/scripts/views/pages/user_dash_wrap.ts
@@ -1,9 +1,0 @@
-import m from 'mithril';
-import { UserDashboard } from './user_dashboard';
-const UserDashWrap = {
-  view: (vnode) => {
-    return m(UserDashboard);
-  },
-};
-
-export default UserDashWrap;
